+++ drivers/css/css.c
@@ -20,7 +20,7 @@
  * the generic setters prevent changing the CSS state, the CSS must always be
  * running.
  */
-uint8_t WEAK
+uint32_t WEAK
 css_get_css_state(void)
 {
 	/* Assume the CSS is always on. */
@@ -31,7 +31,7 @@ css_get_css_state(void)
  * Generic implementation used when no platform support is available. Assume
  * the minimum possible number of clusters is present.
  */
-uint8_t WEAK
+uint32_t WEAK
 css_get_cluster_count(void)
 {
 	/* Assume the CSS contains a single cluster with a single core. */
@@ -43,8 +43,8 @@ css_get_cluster_count(void)
  * the generic setters prevent changing any cluster state, the set of running
  * clusters is always equal to the set of clusters initialized by the boot ROM.
  */
-uint8_t WEAK
-css_get_cluster_state(uint8_t cluster UNUSED)
+uint32_t WEAK
+css_get_cluster_state(uint32_t cluster UNUSED)
 {
 	/* Assume present clusters/cores are always on. */
 	if (cluster >= css_get_cluster_count())
@@ -57,8 +57,8 @@ css_get_cluster_state(uint8_t cluster UNUSED)
  * Generic implementation used when no platform support is available. Assume
  * the minimum possible number of cores is present in each cluster.
  */
-uint8_t WEAK
-css_get_core_count(uint8_t cluster UNUSED)
+uint32_t WEAK
+css_get_core_count(uint32_t cluster UNUSED)
 {
 	/* Assume the CSS contains a single cluster with a single core. */
 	return 1;
@@ -69,8 +69,8 @@ css_get_core_count(uint8_t cluster UNUSED)
  * the generic setters prevent changing any core state, the set of running
  * cores is always equal to the set of cores initialized by the boot ROM.
  */
-uint8_t WEAK
-css_get_core_state(uint8_t cluster, uint8_t core)
+uint32_t WEAK
+css_get_core_state(uint32_t cluster, uint32_t core)
 {
 	/* Assume present clusters/cores are always on. */
 	if (cluster >= css_get_cluster_count())
@@ -87,16 +87,16 @@ css_get_core_state(uint8_t cluster, uint8_t core)
  * lower-level core state API. However, this definition is declared weak for
  * consistency with the other functions and flexibility for future platforms.
  */
-uint8_t WEAK
-css_get_online_cores(uint8_t cluster)
+uint32_t WEAK
+css_get_online_cores(uint32_t cluster)
 {
-	uint8_t cores;
-	uint8_t mask = 0;
+	uint32_t cores;
+	uint32_t mask = 0;
 
 	assert(cluster < css_get_cluster_count());
 
 	cores = css_get_core_count(cluster);
-	for (uint8_t core = 0; core < cores; ++core) {
+	for (uint32_t core = 0; core < cores; ++core) {
 		if (css_get_core_state(cluster, core) != SCPI_CSS_OFF)
 			mask |= BIT(core);
 	}
@@ -119,7 +119,7 @@ css_init(void)
  * to the CSS state by default without a platform-specific implementation.
  */
 int WEAK
-css_set_css_state(uint8_t state)
+css_set_css_state(uint32_t state)
 {
 	/* Reject any attempts to change CSS, cluster, or core power states. */
 	if (state != css_get_css_state())
@@ -134,7 +134,7 @@ css_set_css_state(uint8_t state)
  * to the cluster state by default without a platform-specific implementation.
  */
 int WEAK
-css_set_cluster_state(uint8_t cluster, uint8_t state)
+css_set_cluster_state(uint32_t cluster, uint32_t state)
 {
 	/* Reject any attempts to change CSS, cluster, or core power states. */
 	if (state != css_get_cluster_state(cluster))
@@ -149,7 +149,7 @@ css_set_cluster_state(uint8_t cluster, uint8_t state)
  * to the core state by default without a platform-specific implementation.
  */
 int WEAK
-css_set_core_state(uint8_t cluster, uint8_t core, uint8_t state)
+css_set_core_state(uint32_t cluster, uint32_t core, uint32_t state)
 {
 	/* Reject any attempts to change CSS, cluster, or core power states. */
 	if (state != css_get_core_state(cluster, core))
